implement simons comments
