+++ src/main/java/org/elasticsearch/common/lucene/search/function/ScriptScoreFunction.java
@@ -21,7 +21,6 @@ package org.elasticsearch.common.lucene.search.function;
 
 import org.apache.lucene.index.AtomicReaderContext;
 import org.apache.lucene.search.Explanation;
-import org.elasticsearch.common.lucene.search.function.ScoreCombiner.CombineFunction;
 import org.elasticsearch.script.ExplainableSearchScript;
 import org.elasticsearch.script.SearchScript;
